requirements.txt
asgiref==3.9.0
Django==5.2.4
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
openai==1.93.0
python-dotenv==1.1.1
sqlparse==0.5.3
